refactor(database): move all peewee models to `db_tables.py`
