+++ src/valentina/cogs/characters.py
@@ -9,7 +9,7 @@ from loguru import logger
 from valentina.character.traits import add_trait
 from valentina.character.wizard import CharGenWizard
 from valentina.models.bot import Valentina
-from valentina.models.database import CustomTrait, TraitValue
+from valentina.models.db_tables import CustomTrait, TraitValue
 from valentina.utils.converters import (
     ValidCharacterClass,
     ValidCharacterName,
