+++ src/valentina/cogs/xp.py
@@ -1,5 +1,4 @@
 # mypy: disable-error-code="valid-type"
-
 """A cog for handling XP and leveling up."""
 import discord
 from discord.commands import Option
@@ -8,7 +7,7 @@ from loguru import logger
 
 from valentina.models.bot import Valentina
 from valentina.models.constants import COOL_POINT_VALUE, EmbedColor, XPMultiplier
-from valentina.models.database import CustomTrait, TraitValue
+from valentina.models.db_tables import CustomTrait, TraitValue
 from valentina.utils.converters import ValidCharTrait
 from valentina.utils.helpers import (
     fetch_clan_disciplines,
