+++ src/valentina/models/bot.py
@@ -9,13 +9,12 @@ from aiohttp import ClientSession
 from discord.ext import commands, tasks
 from loguru import logger
 
-from valentina.models import GuildService, MacroService, TraitService, UserService
-from valentina.models.database import DATABASE
+from valentina.models import DatabaseService, GuildService, MacroService, TraitService, UserService
 from valentina.models.database_services import (
     CharacterService,
     ChronicleService,
-    DatabaseService,
 )
+from valentina.models.db_tables import DATABASE
 from valentina.utils import Context
 from valentina.utils.db_backup import DBBackup
