+++ src/valentina/models/macros.py
@@ -3,10 +3,8 @@ from __future__ import annotations
 
 import discord
 from loguru import logger
-from peewee import DateTimeField, DoesNotExist, ForeignKeyField, TextField
 
-from valentina.models.database import BaseModel, CustomTrait, Guild, Trait, User
-from valentina.utils.helpers import time_now
+from valentina.models.db_tables import CustomTrait, Macro, MacroTrait, Trait
 
 
 class MacroService:
@@ -128,54 +126,3 @@ class MacroService:
             logger.debug("CACHE: Purge all macros from cache")
             self._macro_cache.clear()
             self._trait_cache.clear()
-
-
-class Macro(BaseModel):
-    """Macros for quick dice rolls."""
-
-    name = TextField()
-    abbreviation = TextField()
-    description = TextField(null=True)
-    created = DateTimeField(default=time_now)
-    modified = DateTimeField(default=time_now)
-    guild = ForeignKeyField(Guild, backref="macros")
-    user = ForeignKeyField(User, backref="macros")
-
-    def remove(self) -> None:
-        """Delete the macro and associated macro traits."""
-        for mt in self.traits:
-            mt.delete_instance()
-
-        super().delete_instance()
-
-    class Meta:
-        """Meta class for the model."""
-
-        table_name = "macros"
-
-
-class MacroTrait(BaseModel):
-    """Join table for Macro and Trait."""
-
-    macro = ForeignKeyField(Macro, backref="traits")
-    trait = ForeignKeyField(Trait, backref="macros", null=True)
-    custom_trait = ForeignKeyField(CustomTrait, backref="macros", null=True)
-
-    @classmethod
-    def create_from_trait_name(cls, macro: Macro, trait_name: str) -> MacroTrait:
-        """Create a MacroTrait for the given macro and trait_name."""
-        try:
-            trait = Trait.get(Trait.name == trait_name)
-            return cls.create(macro=macro, trait=trait)
-        except DoesNotExist:
-            custom_trait = CustomTrait.get(CustomTrait.name == trait_name)
-            return cls.create(macro=macro, custom_trait=custom_trait)
-
-    class Meta:
-        """Meta class for the model."""
-
-        table_name = "macro_traits"
-        indexes = (
-            (("macro", "trait"), False),
-            (("macro", "custom_trait"), False),
-        )
