+++ src/valentina/models/users.py
@@ -11,7 +11,7 @@ from discord import ApplicationContext
 from loguru import logger
 
 from valentina.models.constants import TraitPermissions, XPPermissions
-from valentina.models.database import Character, GuildUser, User
+from valentina.models.db_tables import Character, GuildUser, User
 from valentina.utils.helpers import time_now
